fix db dsn, models, add migrations

+++ src/models.py
@@ -23,7 +23,7 @@ users = sa.Table(
     sa.Column("last_name", sa.String(256), nullable=False),
     sa.Column("age", sa.Integer, nullable=False),
     sa.Column("sex", sa.Enum(Sex), nullable=False),
-    sa.Column("city_id", sa.Integer, sa.ForeignKey("cities.id"), nullable=False),
+    sa.Column("city_id", sa.BigInteger, sa.ForeignKey("cities.id"), nullable=False),
     sa.Column("created_dt", sa.DateTime, nullable=False, server_default=sa.func.now()),
 )
 
@@ -34,7 +34,7 @@ users_interests = sa.Table(
     sa.Column(
         "id", sa.BigInteger, primary_key=True, autoincrement=True, nullable=False
     ),
-    sa.Column("user_id", sa.Integer, sa.ForeignKey("users.id"), nullable=False),
+    sa.Column("user_id", sa.BigInteger, sa.ForeignKey("users.id"), nullable=False),
     sa.Column("interest", sa.String(256), nullable=False),
 )
 
@@ -45,8 +45,12 @@ users_relations = sa.Table(
     sa.Column(
         "id", sa.BigInteger, primary_key=True, autoincrement=True, nullable=False
     ),
-    sa.Column("first_user_id", sa.Integer, sa.ForeignKey("users.id"), nullable=False),
-    sa.Column("second_user_id", sa.Integer, sa.ForeignKey("users.id"), nullable=False),
+    sa.Column(
+        "first_user_id", sa.BigInteger, sa.ForeignKey("users.id"), nullable=False
+    ),
+    sa.Column(
+        "second_user_id", sa.BigInteger, sa.ForeignKey("users.id"), nullable=False
+    ),
     sa.Column("status", sa.Enum(RelationStatus), nullable=False),
     sa.Column("created_dt", sa.DateTime, nullable=False, server_default=sa.func.now()),
 )
